Bug 1691: configs/common is being abused. This is partly because it's awkward to keep adding things to the org.cougaar.config.path So: code that does ConfigFinder.getInstance() should often now do: ConfigFinder.getInstance(<module>)
where <module> is the name of the module containing the code and
needing this other file -- ie albbn, glm, planning, etc.

The result is that the ConfigFinder will add 4 new directories to the
front of the search path:
$INSTALL/$MOD/configs/$CONFIG
$INSTALL/$MOD/configs
$INSTALL/$MOD/data/$CONFIG
$INSTALL/$MOD/data

where CONFIG is the value of org.cougaar.config -- ie, minitestconfig
for example.

Note that this addition is on a per-module basis - that is, there are
multiple ConfigFinders here.

Next step - split up the files currently in configs/common....