1) Add the WP servlet
2) Replace the old completion servlet with the newer GLM completion servlet

It looks like the 2 completion servlets have different results for
minitestconfig.

The old one puts 3 of 4 Red. The new one shows only 1 as Red
(3-69-ARBN, which has most of its 24 tasks with low or 0 confidence
successful results).

But since the new one shows better completion when it appears done, if
anything this seems an improvement.

(BTW: the old completion servlet claimed to take the /completion as an argument, but in fact appears to
ignore that and always use /completion)