2306:
Workaround for Service ordering problem.

If some Service can't be set, keep the rest of them, on the assumption
that this missing one will show up later.

Since all Services will be set on each pass, the effect is that some
services could in theory be set multiple times.  In the case of
annotated fields, we can check to see if the field already has a value
and skip it if so.  In the case of setter methods, there's nothing to be
done.